Added a __deepcopy__ function to SingleFileReader.
The behviour is to reuse the format instance when the class is copied. This
means that a deep copy operation does not open the HDF5 file again which is
adding lots of overhead. (Having 1 open file handle is probably a better thing
to do anyway!) Everything else is copied as normal. This results in a 1000x
speed improvement in deepcopy of the imageset (in my test from 1s to 0.001s).