Add FormatCBFFullStill as an additional class in FormatCBFFull.
FormatCBFFull cannot handle a full ImageCIF that's a still image, as FormatCBFFull relies on iotbx, which in turns relies on cbflib_adaptbx which in turn expects a goniometer when the image header is read. FormatCBFFullStill derives from FormatStill and FormatCBFFull, overriding get_raw_data to read the data directly.

Single panel only.  FormatCBFMultiTile should read multipanel images, though there may need to be a Still version of that class implemented at some point.