dxtbx file cache: Avoid invalid state when open fails
If the open_method call failed and threw an exception then the cache
is left in an invalid state - the cache_tag matched but the cache object
was the previous, closed object. lazy_file_cache then threw an error the
next time the same file cache was “read”. Adds a test for this issue.

Also: Protect against potentially similar state issues when closing