breaking API considered harmful
Fast_DP relies on the as_str=True behaviour which is now the default.
However when used with old versions of dxtbx Fast_DP cannot count on
this behaviour, so must specify as_str. Fast_DP can't specify the
parameter conditionally as cctbx_project and dxtbx are not properly
versioned and there is no release model. We certainly can't rely on
the new default behaviour as long as old dxtbx versions are around, as
part of DIALS 1.11 (current version of CCP4), DIALS 1.12 (current
release), possibly phenix.

So we will have to carry this around with us for a while, then remove
as_str from fast_dp, only then remove the dummy parameter in dxtbx.