Allow writing of 'None/Null' for scaling models.
This fixes a bug in the cases where the experiment list has a mix
of experiments without and with scaling models, which failed to load
as n_scaling_models < n_experiments. Therefore always make sure that
the list of models is the same length as the list of experiments by
returning None values where appropriate.