Write correct model ids, when null models exist.
Thanks to @ndevenish for discussion on this.
This fixes an error caused when there is a mix of models and None
values. e.g. scaling_models = [None, ScalingModel0], in the output
there is only one scaling model with id 1, which then cannot be
reloaded. It looks like this hasn't been picked up before, as there
are unlikely to have been cases where there is a mix of null and
real models.