Fix the definition of Experiment::is_still
Was incorrectly using the image angular oscillation range rather than the
oscillation rate. Because this only happened if there was a Scan present,
then this wasn't usually a problem - unless a scan is present and the
oscillation happened to end at 0°.

Add a small test for this case.

Fixes #4 and fixes dials/dials#759.