Py 2/3: handle strings in cbflib
cbflib takes and returns Unicode strings in Py3 by default which doesn't work when decoding the raw pixel data using numpy.frombuffer.  So, in an upcoming commit in cctbx_project/cbflib_adaptbx, we'll be adding SWIG_PYTHON_STRICT_BYTE_CHAR to the cbflib compilation, which means in Py3, all the cbflib calls will take and return byte strings.

Because of this, we need to change anywhere we pass raw strings to cbflib such that they are byte strings. We also need to encode any regular strings from Unicode into byte strings before doing string comparisions with cbflib output in python.