Update command libtbx.install_format
This command updates the dxtbx format registry.

There are two possible installation routes:

-g/--global corresponds to a normal pip package installation, and the
format classes will be available to anyone using the dxtbx installation.
Format classes are stored in /build/dxtbx/formats

-u/--user corresponds to a --user pip package installation, and the
format classes will be available to the installing user only, however
they will be available for every python instance for this user.
Format classes are stored in ~/.dxtbx

The command needs to be re-run whenever format classes are added to/
removed from the format class directory.

It can optionally take a path to a local format class file or a URL,
which will then be copied/downloaded into the format class directory
and added to the registry. Existing files will be replaced.