use entry points in format registry (#34)
Instead of using metaclasses to find image format classes we now use
entry points.

A number of related registry lookup methods were deprecated or removed.

A directed acyclical graph is constructed from all registered image
format classes, and traversed for a given image file. We walk down the
DAG if a format class says it understands the file. Multiple formats on
the same level may understand the file, but they must then converge
again into a single node using multiple inheritance.

We no longer eat exceptions from understand() methods. Similarly, when
no matching format was found the datablock find_format methods now
return None and no longer raise exceptions. In both cases the caller
will need to deal with the situation appropriately.

This is part 2 of the work required for #26 and resolves #26.

Format classes are now always evaluated in deterministic order
(alphabetical unless a format hint is specified), which resolves #7.


To update the dxtbx format registry the command libtbx.install_format can be used.
It offers two installation routes:

-g/--global corresponds to a normal pip package installation, and the
format classes will be available to anyone using the dxtbx installation.
Format classes are stored in /build/dxtbx/formats

-u/--user corresponds to a --user pip package installation, and the
format classes will be available to the installing user only, however
they will be available for every python instance for this user.
Format classes are stored in ~/.dxtbx

The command needs to be re-run whenever format classes are added to/
removed from the format class directory.

It can optionally take a path to a local format class file or a URL,
which will then be copied/downloaded into the format class directory
and added to the registry. Existing files will be replaced.