Use more prior information deserializing still ImageSets
This dramatically increases performance when deserialising experiment
lists constructed as one-imageset-per-image, as they currently are for
still image sets. This reduces the load time to around ~5s for an
example 11k image dataset.

Some of the changes here aren't elegant solutions - there are some
severe restrictions imposed by the data model and the cyclic nature of
needing to resolve the format to construct an imageset in order to
deserialize an imageset. These should be confronted at a later date.

- FormatMultiImage.get_imageset no longer attempts to read every model
  for every image if a beam and detector has been provided and a single
  file index has been requested. This prevents the worst of the N² file
  reading loops that was preventing still experiment lists from loading
- FormatMultiImage.get_imageset caches the format instance from the last
  time it was called. If it is immediately asked for the same filename
  and format arguments again, it reuses these instances rather than
  trying to recreate the format class, which did a lot of file IO
- FormatMultiImage also reuses ImageSetData objects from the last time
  it was called - otherwise you would end up with N unique ImageSetData
  objects allocating 8²*N² bytes of pointers (7.3 GB for 11k images!)
- ExperimentListDict._make_stills also caches previous filename-format
  lookups so that this doesn't need to be redone for every image in
  a multi-image file.