Fix runaway memory usage importing stills in FormatMultiImage
If still images are imported as a single-image-per-ImageSet, the Format
class creates an ImageSetData object for every ImageSet. This reserves
pointers to models for every image, which means that for every image
in the file a total of 8*8*N² bytes are reserved, causing runaway
memory usage if a large set of still experiments is loaded non-lazily.