Fix ImageSetData params pickling in Python3
The params dictionary is stored as pickled data on the ImageSetData
object. However, in python3 boost::python appears to automatically
try to decode any std::strings as UTF-8, with apparently no way of
working around this.

Convert the string to a bytes object explicitly for python3 before
passing through to the pickler.

Fixes #57.

Also:
- Clean up implementation of dumps
- Remove excessively complex tuple conversion code