Add FormatPhilLocator and FormatTIFFRayonixSACLA
FormatTIFFRayonixSACLA is an exotic class built for combining TIFFs at SACLA that were recorded on their Rayonix detector with the per-shot wavelength information from the SACLA data stores. It uses a locator file similar to FormatXTC.  The locator is a phil blob that supplies enough metadata to locate the tiff files and match them with the SACLA data scores where the wavelength is stored. FormatTIFFRayonixSACLA allows us to skip the step where the TIFFs are combined with the wavelength into a new hdf5 file.

To do this, FormatTIFFRayonixSACLA derives from four classes, FormatPhilLocator (new class, see below), FormatMultiImage, FormatStill, and FormatTIFFRayonix.  It derives from FormatTIFFRayonix because it uses that class to read the TIFF files. To get around the fact that FormatTIFFRayonix can't understand the locator file, FormatTIFFRayonixSACLA will temporarily modify self.image_file to point to the TIFF in question to read the data, then restore self.image_file to point to the locator.

FormatPhilLocator is a stub class that only understands files of extension .loc that are parseable for phil.  This class was needed as a base class to derive from since FormatMultiImage, FormatStill, and FormatTIFFRayonix all return False for the understand method of the locator file.  I've modified FormatXTC to derive from FormatPhilLocator too.

Note that FormatTIFFRayonixSACLA requires SACLA's dbpy library and can only be used on their interactive nodes (for now).

Co-authored-by: Asmit Bhowmick <asmit3@gmail.com>