Allow compressed FullCBF files to be read (#80)
Previously, miniCBF files could be read from a gzipped/bzipped file, but
fullcbf files could not, because they passed the filename through to
cbflib.

This patch writes some custom cbflib interaction code and uses an API
call that is unexposed via pycbf to pass through a data buffer when a
compressed file is detected. It also reintroduces a limited form of a
test to compress regression image data and verify reading.

Fixes #72.