Fix dials/dials#764 again
The find_rotation obviously only finds rotations, so if none found look down the stack for an axis which is changing and use this. Restores the expected behaviour. Does open the question as to what the oscillation value _should_ be for this since that is now lost and requires local knowledge as to how the beamline is constructed e.g. should be the _bottom_ of the stack in the referenced ticket