Fix panel pedestal (#109)
* Bug fix for panel.pedestal

Calculate the pedestal map if the panel pedestal > 0

* Simplify logic and allow negative pedestal.

There is no need for two boolean flags. Each panel's pedestal must
be checked (so cannot break the loop) and if any pedestal is
significantly non-zero then a full pedestal map must be created.
This also now allows for negative pedestal maps. Fixes #108

* Add newsfragment

* Add explicit test of get_corrected_data

Ensure that providing a gain or pedestal through the panel works
as expected.

* Allow pedestal to override the metadata dictionary

* Replace ad hoc pedestal code by setting panel.pedestal

Instead of subtracting pedestal from get_raw_data (which is the
wrong place), set panel.pedestal appropriately and then rely
on get_corrected_data to subtract the pedestal.