Ensure past.builtins grammar file are generated
before any code is run. When past.builtins is imported the first time
after installation it loads some python grammar file and creates .pickle
file caches. That is all nice and fine. Unless you happen to import it
in a parallel environment the first time, in which case you end up with
nasty race conditions. So ensure the files are generated at
configuration time.

Line can be removed once we drop all references to past.builtins