Format -> FormatFile as a thought experiment
Essentially this is to make room for a new Format base class which makes
no assumptions about the presence or absence of files - the FormatFile
handles anything File related. The higher level one can handle just a string
which could be interpreted as a file path, or a URI, or whatever.

This is me scoping things out. No intention to push any changes to master.