Add -fsigned-char to SConscript
Summit has a set of compilers that has the default as unsigned char instead of signed char, something that is left up to the system in the C++ spec. This causes reading mini cbfs to fail. The solution is to add -fsigned-char to the dxtbx SConscript. This should already be the default on most platforms, but will allow Summit to compile with fewer special flags.

Closes #163