Fix sls eiger 168 (#169)
Include head in dependency tree

Previously we were skipping the first thing that the sample depends on for FormatNeXus - not a problem before because sample depended on sample_x etc. - issue arose following an update to the Eiger file writer to 20.1.16.56035 which now creates files which are recognised as FormatNexus, and the sample depends directly on omega. 

Fixes #168