Support Formats from "scheme://" URL identifiers (#173)
Adds the ability to build Format classes such that
dxtbx.load("scheme://something") will work - e.g. create a format class
from a URL instead of a filename. This also works when passing things in
through dials, e.g. dials.import. This works via checking for a scheme
extracted via urlparse - if a scheme is present then various file-based
handling (open-read cache, absolute expansion) is bypassed.

A Format class must have the schemes it supports explicitly defined via
the schemes attribute:

    def FormatScheme(Format):
        schemes = ["scheme"]

where the default value for schemes is empty (""). If the format class
does not declare support for a scheme, then it will not be called to
understand() the endpoint - this means that no existing format classes
should need to be changed, as they inherit the empty scheme from Format.
This evaluation is done before understand() but after
importing the class.

Resolves #162