Bugfix for hierarchical detectors in NeXus (#175)
Account for 7fde3a8b463dc90422d4e3a45ebe3304f7266a7d, where the first element of the depends on chain is now the input element. Fine for gonio/scan but doesn't work for multipanel detectors, so in dealing with multiple detectors in a hierarchy pop the first element to restore previous behaviour.