Re-write FormatXTCJungfrau to use homogenous coordinates (LCLS Jungfrau only)
Old code used psana's get_pixel_coords to read an array of pixel coordinates in lab space, then np.mean to find the centers of the objects.  This version uses the vectors in the psana calib file directly.

This change also fixes a bug where the parent frame was applied twice at the asic level.  Not noticable in our 1M experiments, but the more complicated 4M showed the bug. This new version uses PSCalib.SegGeometryStore to look up the ASIC layout and gaps in the ASIC frame of reference instead of using the GeoAccess layout, which is in the module frame of reference.

Also rename quad to module as the 1M and 4M at LCLS do not have quadrants in their calib files.