Set the HDF5 plugin path explicitly again (#265)
In 38cc12d, the plugin handling was "fixed" to rely on HDF5 knowing where it's own plugin path was to be found.

However, this is compiled into H5 as an absolute path, so with the way the dials-installer deploys conda environments we've found this is broken for deployment. This reverts to explicitly setting HDF5_PLUGIN_PATH to the new expected path (if it exists). At worst, this is redundant, and with a properly configured conda environment should be completely unnecessary.

Co-authored-by: Markus Gerstel <2102431+Anthchirp@users.noreply.github.com>