Only read pixel_mask if pixel_mask_applied is False
Previously the code would only read the pixel_mask if
pixel_mask_applied was both defined and True. However, if it is
defined as True, then this means the mask has already been applied
to the raw data, therefore we don't need to read the mask in order
to re-apply the mask. Therefore invert the logic.

Convert function masquerading as a class to a function and add tests.