Simplify the return from find_classes
Previously, find_classes had a cumbersome return statement, returning a
list if a single NX_class argument was passed or a tuple of lists if
multiple NX_class arguments were passed.  Instead use find_classes as
the general case, always returning a tuple of lists, and define
find_class as a convenience function for the single-NX_class case,
returning a list.