Escape % in dxtbx.dlsnxs2cbf -h message
Some part of argparse still uses old-style string formatting, which
means that it tries to substitute something for the '%04d' in the help
message of the 'template' argument, which ought to be printed verbatim
when 'dxtbx.dlsnxs2cbf -h' is invoked.  Escaping this with '%%04d' is a
temporary workaround and will break if/when argparse moves to new-style
string formatting, at which point we should revert to '%04d'.