Build against HDF5 shared library on Windows with conda
As https://www.hdfgroup.org/downloads/hdf5/ says:
 > To use the shared libraries on Windows with Visual Studio you must
 > specify the H5_BUILT_AS_DYNAMIC_LIB compile definition

The scons env does not exist at the point where we would like to set
this, hence record a placeholder in env_etc and env.Append later on.
Somewhat nasty, but I don't know an alternative at the moment.

This fixes #329 but there are problems further on down the line.