Properly link to HDF5 shared library on Windows (#330)
- Keep use_conda definitions separate from non use_conda
- Build against HDF5 shared library on Windows with conda

As https://www.hdfgroup.org/downloads/hdf5/ says:
 > To use the shared libraries on Windows with Visual Studio you must
 > specify the H5_BUILT_AS_DYNAMIC_LIB compile definition

In addition, add a more specific failure mode to dataset_as_flex when a
dataset is not valid and some tidying up in SConscript. This gives a
more useful error for debugging than DXTBX_ASSERT(rank == ndims) if the
dataset cannot be accessed.

Fixes #329 but doesn't completely solve the windows-hdf5 problems.