ExperimentListTemplateImporter: deprecated, not removed (#333)
ExperimentListTemplateImporter is deprecated rather than removed, as we found
several modules depending on it. Also, an ExperimentList.from_templates
convenience function has been added to point coders to, instead of the
ExperimentListFactory - we have discussed wanting to move towards this in the
past, so it's better to go straight there rather than forcing people to change
their code twice.

There is no intention at this time of deprecating ExperimentListFactory.