Ignore PermissionError failures to delete temporary files.
This test fails to delete the final copied file (centroid_0009.cbf)
on Windows, "because it is being used by another process" - except
that Resource Monitor says it is the current Python process that
lists it as an associated handle.

Maybe there is a better way around this, but the files are at least
in a temporary directory, so not raising an error when it fails to
delete seems an improvement.