flumpy: Zero-data-copy bridge between flex and numpy (#377)
The new functions flumpy.to_numpy and flumpy.from_numpy in dxtbx.flumpy
allow you to convert flex→numpy and numpy→flex without copying memory.
This means that altering one will alter the other.

to_numpy will convert a python flex object to a numpy array object of
the same dtype, but which is pointing to the internal memory of the flex
array.

from_numpy will convert a numpy object to a flex object matching the
type of the numpy array (if available), but where the flex object is
referencing the internal memory of the numpy object.

The flex arrays are also exposed as a generic PEP3119 buffer buffer
protocol object, `Scuffer`.

Referencing the internal memory means that no memory needs to be copied,
and changes to one object will be reflected in the other. The numpy
array holds a reference to the flex object, ensuring memory lifetime,
and the flex objects will hold an internal reference to the numpy object.

Alongside the standard data types, flex.vec2* and flex.vec3* will be
converted to N×...×2 and N×...×3 numpy arrays, and flex.mat3* will
be converted to N×....×3×3` numpy objects. vec2*, vec3* and mat3*
can also be converted from numpy to flex arrays, but currently require
the use of flumpy.vec_from_numpy (and an N×[2,3] sized numpy array),
or flumpy.mat3_from_numpy (and either a N×9 or N×3×3 numpy array).


The data types that aren't wrapped are flex.std_string and
flex.sym_mat3_double, which don't linearly map to memory.