Add experimental CMake build
This is an experimental implementation of building dxtbx standalone with
the CMake build system.

It will attempt to discover the existing cctbx_project modules through
the CCTBX_BUILD_DIR cache variable, the LIBTBX_BUILD environment
variable, or via python with "import libtbx".

The src/dxtbx/CMakeLists.txt has been kept deliberately simple and
explicit, at the cost of some verbosity.

There is no attempt at this point to integrate with SCons and attempt
to replace the existing approach to general building in a development
environment.