Fix dxtbx.plot_detector_models for matplotlib 3.5
Fixes #475

Thanks @ndevenish for the report and initial starting point at https://github.com/matplotlib/matplotlib/issues/21688.  That implementation didn't give us the nice Z zoom that we want, so after some digging, set_zlim gets us there.  The call to nonsingular is what set_zlim does if zmin == zmax, as is the case for a flat detector.  It seems to expand out the Z a bit. Calling it directly avoids a UserWarning.

Also fix orthographic=True (not sure when this broke)