Add filtering by event code to FormatXTC (#489)
Adds filtering based on codes added to each event during data collection, to support varying triggering modes at LCLS, where some events need to be discarded.  For example, if a laser alternates between an on and off states, the user can add filter.required_present_codes=XXX to the XTC locator file and only accept events with the required codes. Additionally, the user can specify codes that must not be present using filter.required_absent_codes=YYY.

The filter.pre_filter parameter is useful for tasks such as the image viewer where it is important for all events to be removed up front. For example, if a run has 1000 images, but we want to reject based on an event code present in half of them, if pre_filter is True, then get_num_images will report 500. However, this adds a huge performance hit, at the entire XTC stream must be scanned.

For regular processing using dials.stills_process, the default of filter.pre_filter=False is recommended. Because of lazy loading in dxtbx, the image will not be read until the process assigned to it is ready for it. Only then will the filter will be checked.