Return more properties of NX objects with units
The beam centre, count time, and distance properties of NXDetector can
all have units, according to the NeXus standard, but we have previously
only allowed them to be unitless.  Correct that by returning
pint.Quantity instances for each, instead of int | float.

Also use the new nxmx.units function, introduced for the sake of the
above change, to simplify the definition of the temperature property of
NXSample, the frame time and sensor thickness properties of NXDetector,
and the beam size, flux, total flux and wavelength properties of NXBeam.