Check for incorrect SSL_CERT_FILE in dxtbx.install_format
This is normally correctly set by the libtbx dispatchers and rewritten upon
installation. However, on the MacOS .pkg installers, libtbx.refresh is not
run and so SSL_CERT_FILE points to a nonexistent directory.

This patch checks for this case explicitly when installing formats, and
rewrites it to the correct path before attempting to access the download URL.