Use mean oscillation width
If the difference between the axis positions of the first and second
images in a scan is not representative of the average oscillation width
(perhaps because the goniometer is a little slow to reach a steady
speed), then this first oscillation width will not be representative of
the scan as a whole.  Instead, calculate the average oscillation width
and use that to construct the dxtbx.model.Scan instead.