Warn when oscillations differ significantly
In conversation with Christine Beavers (principal beamline scientist of
I15 at Diamond Light Source, formerly of beamline 12.2.2 at the Advanced
Light Source), it seems a deviation in oscillation widths of ≥ 10% from
the median is a sensible threshold for a warning.  Christine describes
an investigation of the stability of the 12.2.2 diffractometer scan
axis:
"What we found is that the jitter on the steps was most dramatic on the
first couple steps, as you expect, so we had a longer run up and down to
achieve a stable velocity.  We essentially threw away a couple images at
the beginning of the scan.  We never saw much more than 2-5% error on
the step width after the lengthened run-up, so 10% should be pretty
safe."