NeXus: use median oscillation width (#526)
Use the median oscillation width instead of the first oscillation width 
when determining the scan axis step size for the whole of a rotation 
scan.  This is particularly helpful when the axis position data are 
read-back values, rather than demand values, and the first few 
oscillations are not representative of the whole scan.  This might occur
if the goniometer is slow to accelerate the demand speed, for example.

Warn when any individual oscillation width differs significantly from
the median.  In conversation with Christine Beavers (principal beamline 
scientist of I15 at Diamond Light Source, formerly of beamline 12.2.2 at
the Advanced Light Source), it seems a deviation in oscillation widths 
of ≥ 10% from the median is a sensible threshold for a warning.  
Christine describes an investigation of the stability of the 12.2.2 
diffractometer scan axis:
"What we found is that the jitter on the steps was most dramatic on the
first couple steps, as you expect, so we had a longer run up and down to
achieve a stable velocity.  We essentially threw away a couple images at
the beginning of the scan.  We never saw much more than 2-5% error on
the step width after the lengthened run-up, so 10% should be pretty
safe."

Co-authored-by: DiamondLightSource-build-server <DiamondLightSource-build-server@users.noreply.github.com>
Co-authored-by: Richard Gildea <rjgildea@users.noreply.github.com>