Update test.
My interpretation of FormatPYunspecified is that the trusted_range
was already defined as [min-good-value,max-good-value], so I left
that definition in place. However, the masking now also assumes
that definition, so the number of masked pixels is now slightly
different. I believe it is now correct, whereas before it was
wrong. I would welcome review from @phyy-nx and @nksauter on this
point.